Move toolbar position logic to toolbar.js
