+++ package.json
@@ -4,7 +4,7 @@
   "description": "Rich Text editor built on top of draft.js",
   "main": "lib/Megadraft.js",
   "dependencies": {
-    "draft-js": "^0.2.2",
+    "draft-js": "^0.4.0",
     "react-portal": "^2.1.0"
   },
   "devDependencies": {
@@ -39,7 +39,7 @@
     "material-ui": "^0.15.0-alpha.2",
     "mocha": "^2.4.5",
     "phantomjs-prebuilt": "^2.1.7",
-    "react": "^0.14.0",
+    "react": "^15.0.0-rc",
     "react-addons-test-utils": "^0.14.7",
     "react-dom": "^0.14.0",
     "react-router": "^2.0.1",
@@ -53,8 +53,8 @@
     "webpack-dev-server": "^1.14.1"
   },
   "peerDependencies": {
-    "react": "^0.14.0",
-    "react-dom": "^0.14.0"
+    "react": "^15.0.0-rc",
+    "react-dom": "^15.0.0-rc"
   },
   "scripts": {
     "start": "webpack-dev-server",
