+++ src/Megadraft.js
@@ -6,7 +6,7 @@
 
 import React, {Component} from "react";
 import ReactDOM from "react-dom";
-import {Editor, RichUtils} from "draft-js";
+import {Editor, RichUtils, getVisibleSelectionRect} from "draft-js";
 
 import Icons from "./icons";
 import Toolbar from "./toolbar";
@@ -41,44 +41,6 @@ export default class Megadraft extends Component {
 
     this.setState({selectedBlock, selectionRange});
     this.props.onChange(editorState);
-  }
-
-  setBarPosition() {
-    const selectionRange = getSelectionRange();
-    const editor = this.refs.editor;
-    const toolbar = ReactDOM.findDOMNode(this.refs.toolbar);
-    const selectionCoords = getSelectionCoords(
-      selectionRange, editor, toolbar);
-
-    if (!this.state.toolbar.position ||
-        this.state.toolbar.position.top !== selectionCoords.offsetTop ||
-        this.state.toolbar.position.left !== selectionCoords.offsetLeft) {
-      this.setState({
-        toolbar: {
-          show: true,
-          position: {
-            top: selectionCoords.offsetTop,
-            left: selectionCoords.offsetLeft
-          }
-        }
-      });
-    }
-  }
-
-  componentDidUpdate() {
-    if (!this.props.editorState.getSelection().isCollapsed()) {
-      if (this.updatingPosition) {
-        clearImmediate(this.updatingPosition);
-      }
-      this.updatingPosition = null ;
-      this.updatingPosition = setImmediate(() => {
-        return this.setBarPosition();
-      });
-    } else {
-      if (this.state.toolbar.show) {
-        this.setState({toolbar: {show: false}});
-      }
-    }
   }
 
   handleKeyCommand(command) {
